Update dependencies: digest and sha2
There are new versions of `digest` and `sha2` available. Update
`pairing-plus` to use the latest stable version for each of these
dependencies.

Note to reviewers:

The necessity for this patch was found while trying to upgrade the
dependencies in a fork of `kzen-curv` [1] (`digest`, `hmac`, `sha2`).

Includes a version bump to v0.20.2, is this correct?

[1] https://github.com/ZenGo-X/curv